Enum оптимизаций перенесен в отдельный файл
